Using webpack watch
