+++ package.json
@@ -11,14 +11,12 @@
     "author": "@tarjeieo, @olemp",
     "main": "lib/index.js",
     "scripts": {
+        "prewatch": "gulp buildJsonResources && gulp buildJsonPreferences",
+        "watch": "concurrently \"webpack --watch\" \"gulp watchUpload\"",
         "buildPnpTemplateFiles": "gulp copyPnpTemplates && gulp localizePnpTemplates && gulp copyResourcesToAssetsTemplate && npm run buildSiteTemplates && gulp copyThirdPartyLibsToTemplate && gulp stampVersionToTemplates",
         "buildSiteTemplates": "gulp buildSiteTemplates --lcid 1044 && gulp buildSiteTemplates --lcid 1033",
         "release": "gulp clean && gulp packageProd && npm run buildPnpTemplateFiles && gulp convertPnpTemplates && gulp copyReleaseFiles && gulp stampVersionToScripts && gulp zipReleasePackage",
         "postversion": "npm run release",
-        "webpack-dev": "webpack --mode development",
-        "watch": "gulp watch",
-        "watch-skipUpload": "gulp watch --skipUpload",
-        "watch-minify": "gulp watch --minify",
         "lint": "eslint --ext .ts,.tsx ./src/js --color --fix"
     },
     "dependencies": {
@@ -38,8 +36,9 @@
         "react-sticky": "^6.0.3"
     },
     "devDependencies": {
-        "@babel/core": "^7.4.0",
-        "@babel/register": "^7.4.0",
+        "@babel/compat-data": "7.8.0",
+        "@babel/core": "7.8.6",
+        "@babel/preset-env": "7.8.0",
         "@types/chai": "3.4.34",
         "@types/es6-promise": "0.0.33",
         "@types/file-saver": "^2.0.0",
@@ -50,20 +49,17 @@
         "@types/react-dom": "16.9.5",
         "@types/sharepoint": "^2016.1.8",
         "@types/sinon": "^5.0.1",
+        "@typescript-eslint/eslint-plugin": "2.30.0",
+        "@typescript-eslint/parser": "2.30.0",
         "ajv": "~5.2.2",
         "array-sort": "^1.0.0",
         "array-unique": "^0.3.2",
         "autoprefixer-stylus": "^0.14.0",
-        "babel-cli": "^6.26.0",
-        "babel-core": "^6.26.3",
-        "babel-loader": "^7.1.4",
-        "babel-plugin-transform-class-properties": "^6.24.1",
-        "babel-plugin-transform-object-assign": "^6.22.0",
-        "babel-plugin-transform-object-rest-spread": "^6.26.0",
-        "babel-preset-env": "^1.7.0",
-        "babel-preset-es2015": "^6.24.1",
-        "babel-preset-react": "^6.24.1",
+        "babel-loader": "8.1.0",
+        "concurrently": "^5.2.0",
         "core-js": "^2.6.3",
+        "eslint": "6.8.0",
+        "eslint-plugin-react": "7.19.0",
         "event-stream": "3.3.4",
         "extend": "^3.0.1",
         "fancy-log": "1.3.3",
@@ -118,14 +114,11 @@
         "typedoc": "^0.14.2",
         "typescript": "3.4.5",
         "uuid": "^3.3.2",
-        "webpack": "^3.12.0",
+        "webpack": "4.42.1",
+        "webpack-cli": "3.3.11",
         "whatwg-fetch": "^3.0.0",
         "xlsx": "^0.14.1",
-        "yargs": "^12.0.5",
-        "eslint": "6.8.0",
-        "eslint-plugin-react": "7.19.0",
-        "@typescript-eslint/eslint-plugin": "2.30.0",
-        "@typescript-eslint/parser": "2.30.0"
+        "yargs": "^12.0.5"
     },
     "peerDependencies": {
         "react": ">=15",
