+++ src/js/Provision/Extensions/index.ts
@@ -27,3 +27,4 @@ export async function ApplyExtensions(context: IProvisionContext): Promise<void>
         throw new ProvisionError(err, 'ApplyExtensions')
     }
 }
+//
